mini-services/egxpy-bridge/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
egxpy>=0.1.0
pydantic>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
